Move from slider to text field for credit hours.
- Remove old slider and switch to using text field
- Migrate from integer to double for credit hours field
- Add migration code
- Refactor controller code to use new field

+++ GradePoint/GPACalculatorViewController.swift
@@ -219,7 +219,7 @@ class GPACalculatorViewController: UIViewController {
         let studentType = StudentType(rawValue: UserDefaults.standard.integer(forKey: userDefaultStudentType))!
         let classes = realm.objects(Class.self)
         var totalPoints: Double = 0.0
-        var totalCreditHours: Int = 0
+        var totalCreditHours: Double = 0.0
 
         /// Calculation of gpa closure block
         let calculate: (Bool) -> Double = { isWeighted in
